Refactor form validations This will stop requiring extra modules to be thrown around and centralizes all form-level validations under the actual form classes.
